refactor!: use NonZeroU32

+++ semilattice-database/src/relation.rs
@@ -90,10 +90,7 @@ impl Database {
         self.relation
             .read()
             .unwrap()
-            .depends(
-                key,
-                &CollectionRow::new(pend_collection_id, pend_row as u32),
-            )
+            .depends(key, &CollectionRow::new(pend_collection_id, pend_row))
             .iter()
             .cloned()
             .collect()
